feat: add shareholders tables

+++ src/database/migrations/20230413195403_delete-amount-product.ts
@@ -1,9 +1,12 @@
 import { Knex } from 'knex'
 
 export async function up(knex: Knex): Promise<void> {
-  return knex.schema.alterTable('product', (table) => {
-    table.dropColumn('amount')
-  })
+  const hasAmountColumn = await knex.schema.hasColumn('product', 'amount')
+  if (hasAmountColumn) {
+    await knex.schema.table('product', (table) => {
+      table.dropColumn('amount')
+    })
+  }
 }
 
 export async function down(knex: Knex): Promise<void> {
